ML_AI_sms_detection_transctions_sms_catagory_fix_V25

+++ app/src/main/java/com/example/expncetracker/exptkr/domain/model/Account.kt
@@ -1,10 +1,11 @@
 package com.example.expncetracker.exptkr.domain.model
 
 import androidx.compose.ui.graphics.vector.ImageVector
+import java.math.BigDecimal
 
 data class Account(
     val name: String,
-    val balance: Double,
+    val balance: BigDecimal,
     val icon: ImageVector,
     val color: androidx.compose.ui.graphics.Color
 )
