+++ app/src/main/java/com/example/expncetracker/exptkr/domain/model/SpendingTrend.kt
@@ -1,6 +1,8 @@
 package com.example.expncetracker.exptkr.domain.model
 
+import java.math.BigDecimal
+
 data class SpendingTrend(
     val label: String, // e.g., "Jan", "Feb"
-    val amount: Double
+    val amount: BigDecimal
 )
