+++ app/src/main/java/com/example/expncetracker/exptkr/ui/accounts/AccountUiModel.kt
@@ -1,10 +1,12 @@
 package com.example.expncetracker.exptkr.ui.accounts
 
+import java.math.BigDecimal
+
 // FIX #5: Screen references this but file was missing
 data class AccountUiModel(
     val id: Long,
     val name: String,
-    val balance: Double,
+    val balance: BigDecimal,
     val type: String,
     val color: Int
 )
